Fixed access_token issue

+++ api/serializers.py
@@ -44,7 +44,7 @@ class CustomerSerializer(serializers.ModelSerializer):
 
 class DynamicVirtualAccountSerializer(serializers.Serializer):
 
-    amount = serializers.DecimalField(max_digits=15, decimal_places=2)
+    amount = serializers.IntegerField(max_digits=15, decimal_places=2)
 
     customer = CustomerSerializer()
